pallet-utility: Fix possible mismatch between native/wasm (#10121)
* pallet-utility: Fix possible mismatch between native/wasm

The `batched_calls_limit` constant value includes the `size_of` of the runtime `Call`. As we compile
the runtime for native/wasm, we need to align the call size to ensure that it is the same on
wasm/native. This also solves the problem of different metadata outputs for the same runtime.

* Review feedback